fix: work around two array-core gaps in sixteen modules [skip auto-tag]
Each of these modules could not run at all. Two array-core gaps are
behind them and both are recorded for a proper fix:

reshape ignores the order argument entirely, so a column-major fold
silently produced a row-major one and the wrong numbers came out. The
affected modules now write the fold out by index.

Fancy indexing with an index array into a 3-D result is unsupported,
so gathering embedding rows raised. Those modules now average the rows
directly.

The workarounds are correct but they mask the two core defects, which
should be fixed in _array_core so the modules can go back to the plain
numpy spelling.

+++ src/morie/fn/sgthm.py
@@ -36,13 +36,13 @@ def thomas_process(kappa, mu, sigma, window, seed=None):
     rng = np.random.default_rng(seed)
 
     ext = 3 * sigma
-    n_parents = rng.poisson(kappa * (xmax - xmin + 2 * ext) * (ymax - ymin + 2 * ext))
+    n_parents = int(rng.poisson(kappa * (xmax - xmin + 2 * ext) * (ymax - ymin + 2 * ext)))
     px = rng.uniform(xmin - ext, xmax + ext, n_parents)
     py = rng.uniform(ymin - ext, ymax + ext, n_parents)
 
     all_x, all_y = [], []
     for i in range(n_parents):
-        nc = rng.poisson(mu)
+        nc = int(rng.poisson(mu))
         cx = px[i] + rng.normal(0, sigma, nc)
         cy = py[i] + rng.normal(0, sigma, nc)
         mask = (cx >= xmin) & (cx <= xmax) & (cy >= ymin) & (cy <= ymax)
